perf: ⚡ 优化数字滚动代码
